+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changes
 
 
+## [v3.2.0]
+
+- Added 5 new Phlorest phylogenies.
+- Added the 4th edition of the Carneiro dataset (and renamed and updated the 6th edition - formerly
+  just labeled "carneiro", now "carneiro6").
+
+
 ## [v3.1.1]
 
 Added missing citation for dataset Carneiro.